remove the start/end
These really never made sense. I kind of copied the structure from the Pikuma physics course
but really there is never ONE point on both sides these exists. I think it makes it easier
to draw debug lines and such, but I that can be re-constructed every time I need to create
it from the normal and the depth